chatbot-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-generativeai==0.8.3
qdrant-client==1.11.2
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
langchain==0.3.0
langchain-community==0.3.0
langchain-google-genai==1.0.3
sentence-transformers==3.1.1
python-multipart==0.0.12
requests==2.32.3